HFP-1449 Fixes setting max height for elements that already has a width
When using setElementHeight the function assumed that the element always would be able to use 100% of the width minus its margins, this is incorrect for elements that already specify a width.
Fixes max height, max width and left position of popup so it does not overflow parent container in height or width, especially important on smaller devices.
Fixes feedback content text conflicting with the close button when feedback is popup, and makes sure contrast of close button is WCAG compliant.